requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.1.0
babel==2.17.0
Beaker==1.12.1
beautifulsoup4==4.13.4
bleach==6.2.0
blivet==3.11.0
blivet-gui==2.6.0
blobfile==3.0.0
Brlapi==0.8.5
cattrs==24.1.2
caugetch==0.0.1
certifi==2025.1.31
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
clipboard==0.0.4
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
crypt_r==3.13.1
cryptography==43.0.0
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
dbus-python==1.3.2
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnf==4.23.0
docstring-to-markdown==0.15
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.18.0
fire==0.7.0
fonttools==4.56.0
fqdn==1.5.1
fros==1.1
fsspec==2025.3.0
getpass4==0.0.14.1
gpg==1.23.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
humanize==4.9.0
i3ipc==2.2.1
idna==3.7
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
iso639==0.1.4
isoduration==20.11.0
jedi==0.19.2
jedi-language-server==0.43.1
Jinja2==3.1.6
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab-vim==4.1.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
langtable==0.0.68
libcomps==0.1.21
libdnf==0.74.0
llama_stack==0.1.7
llama_stack_client==0.1.7
louis==3.33.0
lsprotocol==2023.0.1
lxml==5.3.2
Mako==1.2.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
msgpack==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
nftables==0.1
notebook==7.3.3
notebook_shim==0.2.4
numpy==2.2.3
nvidia-nccl-cu12==2.26.5
nwg-panel==0.9.59
olefile==0.47
opencv-contrib-python==4.11.0.86
openpyxl==3.1.5
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
Paste==3.10.1
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.0.0
platformdirs==4.3.6
ply==3.11
productmd==1.45
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pwquality==1.4.5
pyaml==25.1.0
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.22.0
pycups==2.0.4
pydantic==2.10.6
pydantic_core==2.27.2
pyenchant==3.2.2
pygame==2.6.1
pygls==1.3.1
Pygments==2.19.1
PyGObject==3.48.2
pyinotify==0.9.6
pykickstart==3.58
pyOpenSSL==24.2.1
pyparsing==3.2.2
pyparted==3.13.0
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5_sip==12.15.0
PySocks==1.7.1
pyte==0.8.2
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==3.3.0
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
python-xlib==0.33
pytz==2025.1
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.1
pyynl @ file:///builddir/build/BUILD/kernel-6.14.9-build/kernel-6.14.9/linux-6.14.9-200.fc41.x86_64/tools/net/ynl
pyzmq==26.3.0
rapidfuzz==3.5.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
rpm==4.20.1
scikit-learn==1.6.1
scipy==1.15.2
selinux @ file:///builddir/build/BUILD/libselinux-3.7-build/libselinux-3.7/src
Send2Trash==1.8.3
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.7-build/selinux-3.7/python/sepolicy
setools==4.5.1
setuptools==69.2.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sniffio==1.3.1
sos==4.9.1
soupsieve==2.7
SQLAlchemy==2.0.38
stack-data==0.6.3
systemd-python==235
Tempita==0.5.2
termcolor==2.5.0
terminado==0.18.1
thefuck==3.32
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
torbrowser-launcher==0.3.7
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==1.26.20
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
xgboost==3.0.0
xkbregistry==0.3
